Add consistent way of extracting commands/subcommand names
This was much too prone to manual error.  For now I just went through the
commands that were predominantly 'defaults write' commands as that what
I'm currently the most focused on.

I also removed the duplication of the restarting of processes and made it
happen after any command (which has been given a value) has been run.